Extract VideoRendererEventListener.
