+++ core/src/main/java/com/novoda/noplayer/exoplayer/ExoPlayerTwoFacade.java
@@ -4,37 +4,31 @@ import android.content.Context;
 import android.net.Uri;
 import android.os.Handler;
 import android.os.Looper;
-import android.view.Surface;
 
 import com.google.android.exoplayer2.DefaultLoadControl;
 import com.google.android.exoplayer2.ExoPlayerFactory;
-import com.google.android.exoplayer2.Format;
 import com.google.android.exoplayer2.SimpleExoPlayer;
-import com.google.android.exoplayer2.decoder.DecoderCounters;
 import com.google.android.exoplayer2.source.ExtractorMediaSource;
 import com.google.android.exoplayer2.source.MediaSource;
 import com.google.android.exoplayer2.trackselection.DefaultTrackSelector;
 import com.google.android.exoplayer2.ui.SimpleExoPlayerView;
 import com.google.android.exoplayer2.upstream.DefaultDataSourceFactory;
-import com.google.android.exoplayer2.video.VideoRendererEventListener;
 import com.novoda.noplayer.ContentType;
 
 import java.io.IOException;
 import java.util.List;
 import java.util.concurrent.CopyOnWriteArrayList;
 
-public class ExoPlayerTwoFacade implements VideoRendererEventListener {
+public class ExoPlayerTwoFacade {
 
     private static final boolean RESET_POSITION = true;
     private static final boolean RESET_STATE = false;
-
-    private int videoWidth;
-    private int videoHeight;
 
     private final SimpleExoPlayer exoPlayer;
     private final MediaSourceFactory mediaSourceFactory;
     private final EventListener exoPlayerEventListener;
     private final MediaSourceEventListener mediaSourceEventListener;
+    private final ExoPlayerVideoRendererEventListener videoRendererEventListener;
     private final List<Forwarder> forwarders;
 
     private InternalErrorListener internalErrorListener;
@@ -49,20 +43,30 @@ public class ExoPlayerTwoFacade implements VideoRendererEventListener {
         List<Forwarder> forwarders = new CopyOnWriteArrayList<>();
         EventListener exoPlayerEventListener = new EventListener(forwarders);
         MediaSourceEventListener mediaSourceEventListener = new MediaSourceEventListener(forwarders);
+        ExoPlayerVideoRendererEventListener videoRendererEventListener = new ExoPlayerVideoRendererEventListener(forwarders);
 
-        return new ExoPlayerTwoFacade(exoPlayer, mediaSourceFactory, forwarders, exoPlayerEventListener, mediaSourceEventListener);
+        return new ExoPlayerTwoFacade(
+                exoPlayer,
+                mediaSourceFactory,
+                forwarders,
+                exoPlayerEventListener,
+                mediaSourceEventListener,
+                videoRendererEventListener
+        );
     }
 
     public ExoPlayerTwoFacade(SimpleExoPlayer exoPlayer,
                               MediaSourceFactory mediaSourceFactory,
                               List<Forwarder> forwarders,
                               EventListener exoPlayerEventListener,
-                              MediaSourceEventListener mediaSourceEventListener) {
+                              MediaSourceEventListener mediaSourceEventListener,
+                              ExoPlayerVideoRendererEventListener videoRendererEventListener) {
         this.exoPlayer = exoPlayer;
         this.mediaSourceFactory = mediaSourceFactory;
         this.exoPlayerEventListener = exoPlayerEventListener;
         this.forwarders = forwarders;
         this.mediaSourceEventListener = mediaSourceEventListener;
+        this.videoRendererEventListener = videoRendererEventListener;
     }
 
     public boolean getPlayWhenReady() {
@@ -86,11 +90,11 @@ public class ExoPlayerTwoFacade implements VideoRendererEventListener {
     }
 
     public int getVideoWidth() {
-        return videoWidth;
+        return videoRendererEventListener.videoWidth();
     }
 
     public int getVideoHeight() {
-        return videoHeight;
+        return videoRendererEventListener.videoHeight();
     }
 
     public long getMediaDuration() {
@@ -99,47 +103,6 @@ public class ExoPlayerTwoFacade implements VideoRendererEventListener {
         } else {
             return 0;
         }
-    }
-
-    @Override
-    public void onVideoEnabled(DecoderCounters counters) {
-        // no-op
-    }
-
-    @Override
-    public void onVideoDecoderInitialized(String decoderName, long initializedTimestampMs, long initializationDurationMs) {
-        // no-op
-    }
-
-    @Override
-    public void onVideoInputFormatChanged(Format format) {
-        // no-op
-    }
-
-    @Override
-    public void onDroppedFrames(int count, long elapsedMs) {
-        for (Forwarder forwarder : forwarders) {
-            forwarder.onDroppedFrames(count, elapsedMs);
-        }
-    }
-
-    @Override
-    public void onVideoSizeChanged(int width, int height, int unappliedRotationDegrees, float pixelWidthHeightRatio) {
-        videoHeight = height;
-        videoWidth = width;
-        for (Forwarder forwarder : forwarders) {
-            forwarder.forwardVideoSizeChanged(width, height, unappliedRotationDegrees, pixelWidthHeightRatio);
-        }
-    }
-
-    @Override
-    public void onRenderedFirstFrame(Surface surface) {
-        // no-op
-    }
-
-    @Override
-    public void onVideoDisabled(DecoderCounters counters) {
-        // no-op
     }
 
     public int getBufferedPercentage() {
